+++ src/main/java/dk/kb/kaltura/client/DsKalturaClient.java
@@ -1,9 +1,6 @@
 package dk.kb.kaltura.client;
 
-import com.kaltura.client.enums.ESearchEntryFieldName;
-import com.kaltura.client.enums.ESearchItemType;
-import com.kaltura.client.enums.ESearchOperatorType;
-import com.kaltura.client.enums.MediaType;
+import com.kaltura.client.enums.*;
 import com.kaltura.client.services.BaseEntryService;
 import com.kaltura.client.services.ESearchService;
 import com.kaltura.client.services.MediaService;
@@ -13,6 +10,7 @@ import com.kaltura.client.services.MediaService.RejectMediaBuilder;
 import com.kaltura.client.services.UploadTokenService;
 import com.kaltura.client.types.*;
 import com.kaltura.client.utils.request.BaseRequestBuilder;
+import com.kaltura.client.utils.request.MultiRequestBuilder;
 import com.kaltura.client.utils.response.base.Response;
 import dk.kb.kaltura.enums.FileExtension;
 import dk.kb.kaltura.enums.MimeType;
@@ -21,7 +19,11 @@ import javax.annotation.Nullable;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
-import java.util.*;
+import java.util.Collections;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
 
@@ -35,26 +37,43 @@ import java.util.stream.Collectors;
  */
 public class DsKalturaClient extends DsKalturaClientBase {
 
+    private static final Integer MAX_RETRY_COUNT = 3;
+
+    private final Integer conversionQueueThreshold;
+    private final Integer conversionQueueRetryDelaySeconds;
+
+    private Integer estimatedQueueLength;
+
+
     /**
-     * Instantiate a session to Kaltura that can be used. The sessions can be reused between Kaltura calls without authenticating again.
+     * Instantiate a session to Kaltura that can be used. The sessions can be reused between Kaltura calls without
+     * authenticating again. Either a token/tokenId a adminSecret must be provided for authentication.
      *
-     * @param kalturaUrl              The Kaltura API url. Using the baseUrl will automatic append the API service part to the URL.
-     * @param userId                  The userId that must be defined in the kaltura, userId is email xxx@kb.dk in our kaltura
-     * @param partnerId               The partner id for kaltura. Kind of a collectionId.
-     * @param token                   The application token used for generating client sessions
-     * @param tokenId                 The id of the application token
-     * @param adminSecret             The adminsecret used as password for authenticating. Must not be shared.
-     * @param sessionDurationSeconds  The duration of Kaltura Session in seconds. Beware that when using AppTokens
-     *                                this might have an upper bound tied to the AppToken.
-     * @param sessionRefreshThreshold The threshold in seconds for session renewal.
-     *                                <p>
-     *                                Either a token/tokenId a adminSecret must be provided for authentication.
-     * @throws IOException If session could not be created at Kaltura
+     * @param kalturaUrl                       The Kaltura API url. Using the baseUrl will automatic append the API service part to the URL.
+     * @param userId                           The userId that must be defined in the kaltura, userId is email xxx@kb.dk in our kaltura
+     * @param partnerId                        The partner id for kaltura. Kind of a collectionId.
+     * @param token                            The application token used for generating client sessions
+     * @param tokenId                          The id of the application token
+     * @param adminSecret                      The adminsecret used as password for authenticating. Must not be shared.
+     * @param sessionDurationSeconds           The duration of Kaltura Session in seconds. Beware that when using AppTokens
+     *                                         this might have an upper bound tied to the AppToken.
+     * @param sessionRefreshThreshold          The threshold in seconds for session renewal.
+     * @param conversionQueueThreshold         The threshold for the size of the conversionQueue. If queue gets larger than
+     *                                         this threshold, upload will wait and retry.
+     * @param conversionQueueRetryDelaySeconds the delay in seconds to wait for a retry if the
+     *                                         conversionQueueLength is larger than or equal to
+     *                                         conversionQueueThreshold.
+     * @throws APIException If session could not be created at Kaltura
      */
     public DsKalturaClient(String kalturaUrl, String userId, int partnerId, String token, String tokenId,
-                           String adminSecret, int sessionDurationSeconds, int sessionRefreshThreshold) throws APIException {
+                           String adminSecret, int sessionDurationSeconds, int sessionRefreshThreshold,
+                           int conversionQueueThreshold, int conversionQueueRetryDelaySeconds) throws APIException {
         super(kalturaUrl, userId, partnerId, token, tokenId, adminSecret, sessionDurationSeconds,
                 sessionRefreshThreshold, MAX_BATCH_SIZE);
+        this.conversionQueueThreshold = conversionQueueThreshold;
+        this.conversionQueueRetryDelaySeconds = conversionQueueRetryDelaySeconds;
+
+        estimatedQueueLength = getConversionQueueLength();
     }
 
     /**
@@ -82,7 +101,7 @@ public class DsKalturaClient extends DsKalturaClientBase {
      *
      * @param entryId The unique id in the Kaltura platform for the stream
      * @return True if record was found and deleted. False if the record with the entryId could not be found in Kaltura.
-     * @throws IOException if Kaltura API called failed.
+     * @throws APIException if Kaltura API called failed.
      */
     public boolean deleteStreamByEntryId(String entryId) throws APIException {
         DeleteMediaBuilder request = MediaService.delete(entryId);
@@ -98,12 +117,21 @@ public class DsKalturaClient extends DsKalturaClientBase {
      *
      * @param entryId The unique id in the Kaltura platform for the stream
      * @return True if record was found and blocked. False if the record with the entryId could not be found in Kaltura.
-     * @throws IOException if Kaltura API called failed.
+     * @throws APIException if Kaltura API called failed.
      */
     public boolean blockStreamByEntryId(String entryId) throws APIException {
         RejectMediaBuilder request = MediaService.reject(entryId);
         return buildAndExecute(request, true).isSuccess();
     }
+
+    public ListResponse<MediaEntry> listMediaEntry(MediaEntryFilter filter) throws APIException {
+        return handleRequest(MediaService.list(filter));
+    }
+
+    public int countMediaEntry(MediaEntryFilter filter) throws APIException {
+        return handleRequest(MediaService.count(filter));
+    }
+
 
     /**
      * Search Kaltura for a referenceId. The referenceId was given to Kaltura when uploading the record.<br>
@@ -173,9 +201,9 @@ public class DsKalturaClient extends DsKalturaClientBase {
      * @param kalturaIDs a list of {@code kalturaIDs}.
      * @return a map from {@code kalturaID} to {@code referenceID}.
      * Unresolvable {@code kalturaIDs} will not be present in the map.
-     * @throws IOException if the remote request failed.
+     * @throws APIException if the remote request failed.
      */
-    public Map<String, String> getReferenceIds(List<String> kalturaIDs) throws IOException, APIException {
+    public Map<String, String> getReferenceIds(List<String> kalturaIDs) throws APIException {
         if (kalturaIDs.isEmpty()) {
             log.info("getReferenceIds(kalturaIDs) called with empty list of IDs");
             return Collections.emptyMap();
@@ -196,9 +224,9 @@ public class DsKalturaClient extends DsKalturaClientBase {
      *
      * @param term a search term, such as {@code dr} or {@code tv avisen}.
      * @return a list of Kaltura IDs for matching records, empty if no hits. Max result size is {@link #getBatchSize()}.
-     * @throws IOException if the remote request failed.
+     * @throws APIException if the remote request failed.
      */
-    public List<String> searchTerm(String term) throws IOException, APIException {
+    public List<String> searchTerm(String term) throws APIException {
         ESearchUnifiedItem item = new ESearchUnifiedItem();
         item.setItemType(ESearchItemType.EXACT_MATCH);
         item.searchTerm(term);
@@ -215,7 +243,7 @@ public class DsKalturaClient extends DsKalturaClientBase {
      *
      * @param items at least 1 search item.
      * @return the response from a Kaltura search for the given items.
-     * @throws IOException if the remote request failed.
+     * @throws APIException if the remote request failed.
      */
     @SuppressWarnings("unchecked")
     private Response<ESearchEntryResponse> searchMulti(List<ESearchEntryBaseItem> items) throws APIException {
@@ -283,8 +311,7 @@ public class DsKalturaClient extends DsKalturaClientBase {
      * Create empty upload token to be filled later.
      *
      * @return uploadTokenId of empty uploadToken
-     * @throws IOException
-     * @throws APIException
+     * @throws APIException If request fails
      */
     private String addUploadToken() throws APIException {
         //Get a token that will allow upload
@@ -305,8 +332,7 @@ public class DsKalturaClient extends DsKalturaClientBase {
      * @param uploadTokenId The uploadToken created beforehand
      * @param filePath      The path of file to be uploaded
      * @return The UploadTokenId when upload is complete
-     * @throws IOException
-     * @throws APIException
+     * @throws APIException if request fails
      */
     private String uploadFile(String uploadTokenId, String filePath, MimeType mimeType,
                               String kalturaFileName) throws APIException,
@@ -345,17 +371,18 @@ public class DsKalturaClient extends DsKalturaClientBase {
      * @param referenceId Id external from Kaltura
      * @param tag         Tags on Entry. Used for ease of searching and grouping of entries within KMC.
      * @return EntryId
-     * @throws IOException
-     * @throws APIException
+     * @throws APIException if request fails
      */
     private String addEmptyEntry(MediaType mediaType, String title, String description, String referenceId,
-                                 String tag) throws APIException {
+                                 String tag, Integer conversionProfileId) throws APIException {
         //Create entry with meta data
         MediaEntry entry = new MediaEntry();
         entry.setMediaType(mediaType);
         entry.setName(title);
         entry.setDescription(description);
         entry.setReferenceId(referenceId);
+        entry.setConversionProfileId(conversionProfileId);
+
         if (tag != null) {
             entry.setTags(tag);
         }
@@ -373,18 +400,17 @@ public class DsKalturaClient extends DsKalturaClientBase {
 
 
     /**
-     * Adds content from an uploadToken to an Entry and return flavorID's of that flavor. If FlavorParamID is not null,
+     * Adds content from an uploadToken to an Entry and return entryId. If flavorParamID is not
+     * null,
      * content is added to specified flavor within the Entry. If FlavorParamID is null content is added as source
      * flavor.
      *
      * @param uploadtokenId Upload token with content
      * @param entryId       Entry to receive content
-     * @param flavorParamId Id of the flavorParam where the content needs to be added within the Entry.
      * @return EntryId of updated entry
-     * @throws APIException
-     * @throws IOException
+     * @throws APIException if request fails
      */
-    private String addUploadTokenToEntry(String uploadtokenId, String entryId, Integer flavorParamId)
+    private String addUploadTokenToEntry(String uploadtokenId, String entryId)
             throws APIException {
 
         //Connect uploaded file with meta data entry
@@ -392,14 +418,7 @@ public class DsKalturaClient extends DsKalturaClientBase {
         resource.setToken(uploadtokenId);
         AddContentMediaBuilder requestBuilder;
 
-        if (flavorParamId == null) {
-            requestBuilder = MediaService.addContent(entryId, resource);
-        } else {
-            AssetParamsResourceContainer paramContainer = new AssetParamsResourceContainer();
-            paramContainer.setAssetParamsId(flavorParamId);
-            paramContainer.setResource(resource);
-            requestBuilder = MediaService.addContent(entryId, paramContainer);
-        }
+        requestBuilder = MediaService.addContent(entryId, resource);
 
         try {
             return handleRequest(requestBuilder).getId();
@@ -425,51 +444,23 @@ public class DsKalturaClient extends DsKalturaClientBase {
      * If there for some reason happens an error after the file is uploaded and not connected to the metadata record, it does not
      * seem possible to later see the file in the kaltura administration gui. This error has only happened because I forced it.
      *
-     * @param filePath    File path to the media file to upload.
-     * @param referenceId Use our internal ID's there. This referenceId can be used to find the record at Kaltura and also map to internal KalturaId.
-     * @param mediaType   enum type. MediaType.AUDIO or MediaType.VIDEO
-     * @param title       Name/titel for the resource in Kaltura
-     * @param description , optional description
-     * @param tag         Optional tag. Uploads from the DS should always use tag 'DS-KALTURA'.  There is no backup for this tag in Kaltura and all uploads can be deleted easy.
+     * @param filePath            File path to the media file to upload.
+     * @param referenceId         Use our internal ID's there. This referenceId can be used to find the record at Kaltura and also map to internal KalturaId.
+     * @param mediaType           MediaType.AUDIO or MediaType.VIDEO
+     * @param title               Name/titel for the resource in Kaltura
+     * @param description         Optional description
+     * @param tag                 Optional tag. Uploads from the DS should always use tag 'DS-KALTURA'.  There is no backup for this tag in Kaltura and all uploads can be deleted easy.
+     * @param fileExtension       file extension appended to filename if not present in filepath when uploading to Kaltura
+     * @param conversionProfileId Optional conversionProfileId that match conversion/transcoding profile in Kaltura.
+     *                            If null or 0, the default conversion profile in Kaltura is used.
      * @return The internal id for the Kaltura record. Example format: '0_jqmzfljb'
-     * @throws IOException the io exception
-     */
-    public String uploadMedia(String filePath, String referenceId, MediaType mediaType, String title,
-                              String description,
-                              String tag, FileExtension fileExtension) throws IOException, APIException {
-        return uploadMedia(filePath, referenceId, mediaType, title, description, tag, null, fileExtension);
-    }
-
-    /**
-     * Upload a video or audio file to Kaltura.
-     * The upload require 4 API calls to Kaltura
-     * <p><ul>
-     * <li> Request a upload token
-     * <li> Upload file using the upload token. Get a tokenID for the upload
-     * <li> Create the metadata record in Kaltura
-     * <li> Connect the metadata record with the tokenID
-     * </ul><p>
-     * <p>
-     * </ul><p>
-     * <p>
-     * If there for some reason happens an error after the file is uploaded and not connected to the metadata record, it does not
-     * seem possible to later see the file in the kaltura administration gui. This error has only happened because I forced it.
-     *
-     * @param filePath      File path to the media file to upload.
-     * @param referenceId   Use our internal ID's there. This referenceId can be used to find the record at Kaltura and also map to internal KalturaId.
-     * @param mediaType     enum type. MediaType.AUDIO or MediaType.VIDEO
-     * @param title         Name/titel for the resource in Kaltura
-     * @param description   , optional description
-     * @param tag           Optional tag. Uploads from the DS should always use tag 'DS-KALTURA'.  There is no backup for this tag in Kaltura and all uploads can be deleted easy.
-     * @param flavorParamId Optional flavorParamId. This sets what flavor the file should be uploaded as. If not set flavor
-     *         will be source, i.e. flavorParamId = 0.
-     * @param fileExt fileextension appended to filename if not present in filepath when uploading to Kaltura
-     * @return The internal id for the Kaltura record. Example format: '0_jqmzfljb'
-     * @throws IOException the io exception
+     * @throws IOException  the io exception
+     * @throws APIException Thrown when API request goes wrong. Message only readable when using java 11, since
+     *                      implementation uses reflection.
      */
     public String uploadMedia(String filePath, String referenceId, MediaType mediaType,
-                              String title, String description, String tag, @Nullable Integer flavorParamId,
-                              FileExtension fileExt)
+                              String title, String description, String tag,
+                              FileExtension fileExtension, @Nullable Integer conversionProfileId)
             throws IOException, APIException {
 
         if (referenceId == null) {
@@ -479,21 +470,93 @@ public class DsKalturaClient extends DsKalturaClientBase {
             throw new IllegalArgumentException("Kaltura mediaType must be defined");
         }
 
-        if (fileExt == null) {
-            throw new IllegalArgumentException("fileExt must be defined");
+        if (fileExtension == null) {
+            throw new IllegalArgumentException("fileExtension must be defined");
         }
 
-        FileExtension.checkExtension(filePath, fileExt);
+        conversionQueueCheckAndWait();
 
-        MimeType mimeType = MimeType.fromFileExtension(fileExt);
-        String kalturaFileName = referenceId + fileExt.getExtension();
+        FileExtension.checkExtension(filePath, fileExtension);
+
+        MimeType mimeType = MimeType.fromFileExtension(fileExtension);
+        String kalturaFileName = referenceId + fileExtension.getExtension();
 
         String uploadTokenId = addUploadToken();
         uploadFile(uploadTokenId, filePath, mimeType, kalturaFileName);
-        String entryId = addEmptyEntry(mediaType, title, description, referenceId, tag);
-        addUploadTokenToEntry(uploadTokenId, entryId, flavorParamId);
-
+        String entryId = addEmptyEntry(mediaType, title, description, referenceId, tag, conversionProfileId);
+        addUploadTokenToEntry(uploadTokenId, entryId);
+        estimatedQueueLength++; // Add 1 to conversion queue
         return entryId;
     }
+
+
+    /**
+     * Checks the conversionQueue and waits if too long. This method first looks at the estimated queue and only
+     * checks the actual queue length if the estimated queue surpasses the conversionQueueThreshold. If so, it
+     * updates the estimatedQueueLength with the actual queueLength.
+     *
+     * @throws APIException     If Api request fails.
+     * @throws RuntimeException is thrown if the queue length keeps being full after {@value MAX_RETRY_COUNT}
+     */
+    private void conversionQueueCheckAndWait() throws APIException {
+        if (estimatedQueueLength < conversionQueueThreshold) {
+            return;
+        }
+
+        int retryCount = 1;
+        while (retryCount <= MAX_RETRY_COUNT) {
+            estimatedQueueLength = getConversionQueueLength(); // update estimated with actual queue length from Kaltura
+            if (estimatedQueueLength < conversionQueueThreshold) {
+                return;
+            }
+
+            log.info("Kaltura Conversion Queue (conversionQueueLength: {}), larger than threshold"
+                            + "(conversionQueueThreshold: {}), retry {} in {} seconds",
+                    estimatedQueueLength,
+                    conversionQueueThreshold,
+                    retryCount,
+                    conversionQueueRetryDelaySeconds);
+
+            try {
+                TimeUnit.SECONDS.sleep(conversionQueueRetryDelaySeconds);
+            } catch (InterruptedException ie) {
+                log.error("Thread was interrupted while waiting for Kaltura conversion queue to decrease");
+                Thread.currentThread().interrupt();
+            }
+
+            retryCount++;
+        }
+
+        throw new RuntimeException("Maximum retry count (" + MAX_RETRY_COUNT + ") was reached while waiting for " +
+                "conversion queue");
+    }
+
+    /**
+     * Get the length of the conversion queue from Kaltura. We define the conversion queue as the count of all
+     * entries with status or replacementStatus that indicates that it is waiting for or in
+     * progress of being converted/transcoded.
+     * <p>
+     * This method uses multiRequest, which essentially does multiple API requests with one http request,
+     * reducing Round-trip time significantly.
+     * @return length of current queue
+     * @throws APIException Thrown when API request fails.
+     */
+    private int getConversionQueueLength() throws APIException {
+
+        MediaEntryFilter replacementFilter = new MediaEntryFilter();
+        replacementFilter.setReplacementStatusIn(EntryReplacementStatus.APPROVED_BUT_NOT_READY.getValue()
+                + "," + EntryReplacementStatus.READY_BUT_NOT_APPROVED.getValue() + "," + EntryReplacementStatus.NOT_READY_AND_NOT_APPROVED);
+
+        MediaEntryFilter conversionFilter = new MediaEntryFilter();
+        conversionFilter.setStatusIn(EntryStatus.IMPORT.getValue() + "," + EntryStatus.PRECONVERT.getValue());
+
+        MultiRequestBuilder multiRequestBuilder =
+                MediaService.count(replacementFilter).add(MediaService.count(conversionFilter));
+        List<Object> countList = handleRequest(multiRequestBuilder);
+        int sum = countList.stream().mapToInt(x -> (int) x).sum();
+        log.debug("Queue length is : {}", sum);
+        return sum;
+    }
+
 
 }