+++ src/main/java/dk/kb/kaltura/client/DsKalturaClientBase.java
@@ -46,12 +46,12 @@ public abstract class DsKalturaClientBase {
     private final String adminSecret;
     private final int sessionKeepAliveSeconds;
     private long lastSessionStart = 0;
-    private final int sessionRefreshThreshold;
     private final int sessionDurationSeconds;
     private int batchSize;
 
     /**
-     * Instantiate a session to Kaltura that can be used. The sessions can be reused between Kaltura calls without authenticating again.
+     * Instantiate a session to Kaltura that can be used. The sessions can be reused between Kaltura calls without
+     * authenticating again. Either a token/tokenId a adminSecret must be provided for authentication.
      *
      * @param kalturaUrl              The Kaltura API url. Using the baseUrl will automatic append the API service part to the URL.
      * @param userId                  The userId that must be defined in the kaltura, userId is email xxx@kb.dk in our kaltura
@@ -62,14 +62,11 @@ public abstract class DsKalturaClientBase {
      * @param sessionDurationSeconds  The duration of Kaltura Session in seconds. Beware that when using AppTokens
      *                                this might have an upper bound tied to the AppToken.
      * @param sessionRefreshThreshold The threshold in seconds for session renewal.
-     *                                <p>
-     *                                Either a token/tokenId a adminSecret must be provided for authentication.
      */
     public DsKalturaClientBase(String kalturaUrl, String userId, int partnerId, String token, String tokenId,
                                String adminSecret, int sessionDurationSeconds, int sessionRefreshThreshold,
                                int batchSize) throws APIException {
         this.sessionDurationSeconds = sessionDurationSeconds;
-        this.sessionRefreshThreshold = sessionRefreshThreshold;
         this.sessionKeepAliveSeconds = sessionDurationSeconds - sessionRefreshThreshold;
         this.kalturaUrl = kalturaUrl;
         this.userId = userId;
@@ -128,7 +125,7 @@ public abstract class DsKalturaClientBase {
      * @return the result of the executed request
      * @throws APIException if an API error occurs during the request execution
      */
-    protected <ReturnedType, SelfType extends BaseRequestBuilder<ReturnedType, SelfType>>
+    public <ReturnedType, SelfType extends BaseRequestBuilder<ReturnedType, SelfType>>
     ReturnedType handleRequest(SelfType requestBuilder) throws APIException {
         return handleRequest(requestBuilder, true);
     }