+++ src/test/java/dk/kb/kaltura/KalturaApiIntegrationTest.java
@@ -1,7 +1,9 @@
 package dk.kb.kaltura;
 
+import com.kaltura.client.enums.EntryStatus;
 import com.kaltura.client.enums.MediaType;
 import com.kaltura.client.types.APIException;
+import com.kaltura.client.types.MediaEntryFilter;
 import dk.kb.kaltura.client.DsKalturaClient;
 import dk.kb.kaltura.config.ServiceConfig;
 import dk.kb.kaltura.enums.FileExtension;
@@ -13,7 +15,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.Arrays;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -131,7 +133,7 @@ public class KalturaApiIntegrationTest {
     }
 
     @Test
-    public void simpleSearch() throws IOException, APIException {
+    public void simpleSearch() throws APIException {
         List<String> ids = getClient().searchTerm("dr");
         assertFalse(ids.isEmpty(), "Search result should not be empty");
         System.out.println(ids);
@@ -158,7 +160,7 @@ public class KalturaApiIntegrationTest {
      *
      */
     @Test
-    public void kalturaUpload() throws Exception {
+    public void uploadDefault() throws Exception {
         DsKalturaClient clientSession = getClient();
         String file = "/home/xxxx/Videos/test1.mp4"; // <-- Change to local video file
         String referenceId = "ref_test_1234s";
@@ -167,78 +169,81 @@ public class KalturaApiIntegrationTest {
         String title = "test2 title from unittest";
         String description = "test2 description from unittest";
         FileExtension fileExtension = FileExtension.MP4;
-        String kalturaId = clientSession.uploadMedia(file, referenceId, mediaType, title, description, tag, fileExtension);
-        assertNotNull(kalturaId);
-    }
+        int conversionProfileId = 0; // <-- change to relevant conversionProfileId found in KMC
 
-    /**
-     * When uploading a file to Kaltura, remember to delete it from the Kaltura
-     *
-     */
-    @Test
-    public void kalturaUploadWithFlavorParam() throws Exception {
-        DsKalturaClient clientSession = getClient();
-        String file = "/home/adpe/IdeaProjects/ds-parent/ds-kaltura/src/test/resources/test_files/goodVideo3.mp4"; // <--
-        // Change to local video file
-        String referenceId = "ref_test_1234s";
-        MediaType mediaType = MediaType.VIDEO;
-        String tag = "DS-KALTURA"; //This tag is use for all upload from DS to Kaltura
-        String title = "test3 title from unittest";
-        String description = "test3 description from unittest";
-        Integer flavorParamId = 3; // <-- Change according to MediaType. 3 for lowQ video and 359 for audio
-        FileExtension fileExt = FileExtension.MP4;
+        //Upload with default flavor and default conversionProfileID
         String kalturaId = clientSession.uploadMedia(file, referenceId, mediaType, title, description, tag,
-                flavorParamId, fileExt);
+                fileExtension, conversionProfileId);
         assertNotNull(kalturaId);
     }
 
-    @Test
-    public void kalturaUploadWithExtension() throws Exception {
-        DsKalturaClient clientSession = getClient();
-        String file = "/home/adpe/IdeaProjects/ds-parent/ds-kaltura/src/test/resources/test_files/goodVideo2.MP4"; // <-- Change to local video file with file extension
-        String referenceId = "ref_test_1234s";
-        MediaType mediaType = MediaType.VIDEO;
-        String tag = "DS-KALTURA"; //This tag is use for all upload from DS to Kaltura
-        String title = "test3 title from unittest";
-        String description = "test3 description from unittest";
-        Integer flavorParamId = 3; // <-- Change according to MediaType. 3 for lowQ video and 359 for audio
-        FileExtension fileExt = FileExtension.MP4;
-        String kalturaId = clientSession.uploadMedia(file, referenceId, mediaType, title, description, tag,
-                flavorParamId, fileExt);
-        assertNotNull(kalturaId);
-    }
+
 
     @Test
-    public void kalturaUploadNoex() throws Exception {
+    public void uploadMisMatchExt() throws Exception {
         DsKalturaClient clientSession = getClient();
-        String file = "/home/adpe/IdeaProjects/ds-parent/ds-kaltura/src/test/resources/test_files/goodAudio"; // <--
-        // Change to local video file without extension
+        String file = "/path/to/file"; //<--Change to local video file with mp4
         String referenceId = "ref_test_1234s";
         MediaType mediaType = MediaType.AUDIO;
         String tag = "DS-KALTURA"; //This tag is use for all upload from DS to Kaltura
         String title = "test3 title from unittest";
         String description = "test3 description from unittest";
-        Integer flavorParamId = 359; // <-- Change according to MediaType. 3 for lowQ video and 359 for audio
-        String kalturaId = clientSession.uploadMedia(file, referenceId, mediaType, title,
-                description, tag, flavorParamId, FileExtension.MP3);
-        assertNotNull(kalturaId);
-    }
+        int conversionProfileId = 0;
 
-
-    @Test
-    public void kalturaUploadMisMatchExt() throws Exception {
-        DsKalturaClient clientSession = getClient();
-        String file = "/home/adpe/IdeaProjects/ds-parent/ds-kaltura/src/test/resources/test_files/goodVideo3.mp4"; //
-        // <--Change to local video file with mp4
-        String referenceId = "ref_test_1234s";
-        MediaType mediaType = MediaType.AUDIO;
-        String tag = "DS-KALTURA"; //This tag is use for all upload from DS to Kaltura
-        String title = "test3 title from unittest";
-        String description = "test3 description from unittest";
-        Integer flavorParamId = 359;// <-- Change according to MediaType. 3 for lowQ video and 359 for audio
         Throwable t = assertThrows(Exception.class, () -> clientSession.uploadMedia(file, referenceId, mediaType, title,
-                description, tag, flavorParamId, FileExtension.MP3));
+                description, tag, FileExtension.MP3, conversionProfileId));
         log.debug(t.toString());
+    }
+
+
+    @Test
+    public void uploadBatchAndCleanUp() throws Exception {
+        int uploadCount = 100;
+
+        DsKalturaClient clientSession = getClient();
+        List<String> entryIdList = new ArrayList<>();
+
+        long startUploadTime = System.currentTimeMillis();
+
+        for (int i = 0; i < uploadCount; i++) {
+
+            String file = "/home/adpe/IdeaProjects/ds-parent/ds-kaltura/src/test/resources/test_files/goodAudio2.noex"; //<-- Change to local video file with mp4
+            String referenceId = "ref_test_" + i;
+            MediaType mediaType = MediaType.AUDIO;
+            String tag = "TEST-TRANSCODING-SPEED"; //This tag is use for all upload from DS to Kaltura
+            String title = "conversion test : " + i;
+            String description = "test3 description from unittest";
+            FileExtension fileExtension = FileExtension.MP3;
+            int conversionProfileId = 1507;
+
+            //Upload to specified conversionprofile and default flavor (Source)
+            String kalturaId = clientSession.uploadMedia(file, referenceId, mediaType, title, description, tag,
+                    fileExtension, conversionProfileId);
+            entryIdList.add(kalturaId);
+        }
+
+        long endUploadTime = System.currentTimeMillis();
+
+        // Check if done transcoding
+        MediaEntryFilter filter = new MediaEntryFilter();
+        filter.setIdIn(String.join(",", entryIdList));
+        filter.setStatusNotEqual(EntryStatus.READY);
+        int done = uploadCount;
+        while (done > 0) {
+            done = clientSession.countMediaEntry(filter);
+            log.info("Converting count: {}", done);
+            Thread.sleep(5000);
+        }
+
+        //Log Time
+        long convertedTime = System.currentTimeMillis();
+        log.info("Time to upload: {}", endUploadTime - startUploadTime);
+        log.info("Time to upload and convert: {}", convertedTime - startUploadTime);
+
+        //Clean up
+        for (String entry : entryIdList) {
+            clientSession.deleteStreamByEntryId(entry);
+        }
     }
 
     private DsKalturaClient getClient() throws APIException {
@@ -251,7 +256,9 @@ public class KalturaApiIntegrationTest {
                 conf.getString("tokenId"),
                 conf.getString("adminSecret", null),
                 conf.getInteger("sessionDurationSeconds", DEFAULT_SESSION_DURATION_SECONDS),
-                conf.getInteger("sessionRefreshThreshold", DEFAULT_REFRESH_THRESHOLD));
+                conf.getInteger("sessionRefreshThreshold", DEFAULT_REFRESH_THRESHOLD),
+                conf.getInteger("conversionQueueThreshold"),
+                conf.getInteger("conversionQueueDelaySeconds"));
     }
 
 }